Organize connectivity module
