+++ rustworkx/rustworkx.pyi
@@ -17,7 +17,8 @@ from .digraph import PyDiGraph as PyDiGraph
 
 from typing import Optional, Set, List, Dict, TypeVar, Tuple, Callable, Union
 
-from .centrality import  *
+from .centrality import *
+from .connectivity import *
 from .isomorphism import *
 from .layout import *
 from .link_analysis import *
@@ -39,11 +40,6 @@ class NegativeCycle(Exception): ...
 class JSONSerializationError(Exception): ...
 class FailedToConverge(Exception): ...
 
-def articulation_points(graph: PyGraph, /) -> Set[int]: ...
-def biconnected_components(graph: PyGraph, /) -> BiconnectedComponents: ...
-def chain_decomposition(graph: PyGraph, /, source: Optional[int] = ...) -> Chains: ...
-def connected_components(graph: PyGraph, /) -> List[Set[int]]: ...
-def cycle_basis(graph: PyGraph, /, root: Optional[int] = ...) -> List[List[int]]: ...
 def collect_runs(
     graph: PyDiGraph[_S, _T],
     filter_fn: Callable[[_S], bool],
@@ -69,22 +65,6 @@ def dag_weighted_longest_path_length(
     weight_fn: Callable[[int, int, _T], float],
     /,
 ) -> float: ...
-def digraph_adjacency_matrix(
-    graph: PyDiGraph[_S, _T],
-    /,
-    weight_fn: Optional[Callable[[_T], float]] = ...,
-    default_weight: float = ...,
-    null_value: float = ...,
-    parallel_edge: str = ...,
-) -> np.ndarray: ...
-def graph_adjacency_matrix(
-    graph: PyGraph[_S, _T],
-    /,
-    weight_fn: Optional[Callable[[_T], float]] = ...,
-    default_weight: float = ...,
-    null_value: float = ...,
-    parallel_edge: str = ...,
-) -> np.ndarray: ...
 def digraph_core_number(
     graph: PyDiGraph,
     /,
@@ -156,14 +136,7 @@ def graph_union(
     merge_edges: bool = ...,
 ) -> PyGraph[_S, _T]: ...
 def is_planar(graph: PyGraph, /) -> bool: ...
-def is_connected(graph: PyGraph, /) -> bool: ...
 def is_directed_acyclic_graph(graph: PyDiGraph, /) -> bool: ...
-def is_weakly_connected(graph: PyDiGraph, /) -> bool: ...
-def number_connected_components(graph: PyGraph, /) -> int: ...
-def number_weakly_connected_components(graph: PyDiGraph, /) -> bool: ...
-def node_connected_component(graph: PyGraph, node: int, /) -> Set[int]: ...
-def strongly_connected_components(graph: PyDiGraph, /) -> List[List[int]]: ...
-def weakly_connected_components(graph: PyDiGraph, /) -> List[Set[int]]: ...
 def topological_sort(graph: PyDiGraph, /) -> NodeIndices: ...
 def lexicographical_topological_sort(
     dag: PyDiGraph[_S, _T],
