+++ rustworkx/tree.pyi
